+++ src/main/java/org/folio/processing/events/services/processor/EventProcessorImpl.java
@@ -4,7 +4,7 @@ import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import org.folio.DataImportEventPayload;
 import org.folio.processing.events.services.handler.EventHandler;
-import org.folio.processing.exceptions.EventProcessingException;
+import org.folio.processing.exceptions.EventHandlerNotFoundException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,6 +12,8 @@ import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
 import static java.lang.String.format;
+import static org.folio.processing.events.EventManager.POST_PROCESSING_INDICATOR;
+import static org.folio.processing.events.EventManager.POST_PROCESSING_RESULT_EVENT_KEY;
 
 public class EventProcessorImpl implements EventProcessor {
 
@@ -30,21 +32,31 @@ public class EventProcessorImpl implements EventProcessor {
         EventHandler eventHandler = optionalEventHandler.get();
         String eventType = eventPayload.getEventType();
         long startTime = System.nanoTime();
-        eventHandler.handle(eventPayload).whenComplete((payload, throwable) -> {
-          logEventProcessingTime(eventType, startTime, eventPayload);
-          if (throwable != null) {
-            future.completeExceptionally(throwable);
-          } else {
-            future.complete(payload);
-          }
-        });
+        eventHandler.handle(eventPayload)
+          .thenApply(dataImportEventPayload -> eventHandler.isPostProcessingNeeded() ? preparePayloadForPostProcessing(dataImportEventPayload, eventHandler) : dataImportEventPayload)
+          .whenComplete((payload, throwable) -> {
+            logEventProcessingTime(eventType, startTime, eventPayload);
+            if (throwable != null) {
+              future.completeExceptionally(throwable);
+            } else {
+              future.complete(payload);
+            }
+          });
       } else {
-        future.completeExceptionally(new EventProcessingException(format("No suitable handler found for %s event type", eventPayload.getEventType())));
+        LOG.warn("No suitable handler found for {} event type and current profile {}", eventPayload.getEventType(), eventPayload.getCurrentNode().getContentType());
+        future.completeExceptionally(new EventHandlerNotFoundException(format("No suitable handler found for %s event type", eventPayload.getEventType())));
       }
     } catch (Exception e) {
       future.completeExceptionally(e);
     }
     return future;
+  }
+
+  private DataImportEventPayload preparePayloadForPostProcessing(DataImportEventPayload dataImportEventPayload, EventHandler eventHandler) {
+    dataImportEventPayload.getContext().put(POST_PROCESSING_INDICATOR, Boolean.toString(true));
+    dataImportEventPayload.getContext().put(POST_PROCESSING_RESULT_EVENT_KEY, dataImportEventPayload.getEventType());
+    dataImportEventPayload.setEventType(eventHandler.getPostProcessingInitializationEventType());
+    return dataImportEventPayload;
   }
 
   @Override