+++ src/main/java/org/folio/processing/exceptions/EventHandlerNotFoundException.java
@@ -1,0 +1,8 @@
+package org.folio.processing.exceptions;
+
+public class EventHandlerNotFoundException extends EventProcessingException {
+
+  public EventHandlerNotFoundException(String message) {
+    super(message);
+  }
+}