+++ src/test/java/org/folio/processing/events/handlers/InstancePostProcessingEventHandler.java
@@ -1,0 +1,33 @@
+package org.folio.processing.events.handlers;
+
+import io.vertx.core.json.JsonObject;
+import org.folio.ActionProfile;
+import org.folio.DataImportEventPayload;
+import org.folio.processing.events.services.handler.EventHandler;
+
+import java.util.concurrent.CompletableFuture;
+
+import static org.folio.ActionProfile.FolioRecord.INSTANCE;
+import static org.folio.DataImportEventTypes.DI_INVENTORY_INSTANCE_UPDATED;
+import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.ACTION_PROFILE;
+
+public class InstancePostProcessingEventHandler implements EventHandler {
+
+  public static final String POST_PROCESSING_RESULT_EVENT = DI_INVENTORY_INSTANCE_UPDATED.value();
+
+  @Override
+  public CompletableFuture<DataImportEventPayload> handle(DataImportEventPayload eventPayload) {
+    eventPayload.getEventsChain().add(eventPayload.getEventType());
+    return CompletableFuture.completedFuture(eventPayload);
+  }
+
+  @Override
+  public boolean isEligible(DataImportEventPayload eventPayload) {
+    if (eventPayload.getCurrentNode() != null && ACTION_PROFILE == eventPayload.getCurrentNode().getContentType()) {
+      ActionProfile actionProfile = JsonObject.mapFrom(eventPayload.getCurrentNode().getContent()).mapTo(ActionProfile.class);
+      return actionProfile.getFolioRecord() == INSTANCE;
+    }
+    return false;
+  }
+
+}